Agent/src/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
mysql-connector-python==8.3.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.6.1
httpx==0.27.2
tiktoken==0.5.2

# AI and ML
openai==1.13.3
sentencepiece==0.1.99
google-generativeai==0.3.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.5

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Additional dependencies
beautifulsoup4==4.12.3
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
python-multipart==0.0.9
coverage==7.4.1 